moved the app.py to a flask specific folder
